requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pypdf
tiktoken
groq
httpx
python-dotenv
tqdm